refactor(review): extract CiStatusEvaluator from ReviewOrchestrator (#250)
Step 1 of the ReviewOrchestrator decomposition: move the self-contained
CI-status subsystem into its own @ApplicationScoped collaborator.

Pure extract-class, no behaviour change. Moved verbatim:
resolveRequiredContexts (+ rulesets/classic-protection helpers),
evaluateCiChecks, collectPaged, the offending-check/status/missing
collectors, classifyCheckRun/classifyStatus, and the bot-check filter —
plus the CI constants they own. ReviewOrchestrator injects the evaluator
and delegates the two entry points; it keeps CI_PENDING for the
no-issues review body. The #217 unknown-state-holds-approval and #236
ruleset-read behaviours move with the code.

Tests: the EvaluateCiChecks and ResolveRequiredContexts cases move to a
new CiStatusEvaluatorTest (24 cases, exercising the evaluator directly
with mocked clients); ReviewOrchestratorTest keeps CiGatingThroughReview
and CiChecksGating as integration coverage, wiring a real
CiStatusEvaluator built from the mocked clients. Full suite green,
spotbugs + spotless clean.
