+++ src/main/java/dev/thiagogonzaga/thrillhousebot/review/ReviewOrchestrator.java
@@ -35,14 +35,9 @@ import java.time.Instant;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Optional;
-import java.util.Set;
 import java.util.function.Consumer;
-import java.util.function.IntFunction;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 
 @ApplicationScoped
@@ -65,21 +60,8 @@ public class ReviewOrchestrator {
   private static final String EVENT_REQUEST_CHANGES = "REQUEST_CHANGES";
   private static final String EVENT_COMMENT = "COMMENT";
 
-  // CI check evaluation constants
-  private static final String CI_SUCCESS = "success";
+  // CI check status shown in the no-issues review body; CI evaluation lives in CiStatusEvaluator.
   private static final String CI_PENDING = "pending";
-  private static final String CI_FAILING = "failing";
-
-  // Lowercased token identifying ThrillhouseBot's own checks/apps among CI results.
-  private static final String THRILLHOUSEBOT_TOKEN = "thrillhousebot";
-
-  // Check-run conclusions that count as a passing check (everything else is a failure)
-  private static final Set<String> PASSING_CONCLUSIONS = Set.of(CI_SUCCESS, "skipped", "neutral");
-
-  // CI status pagination: GitHub serves at most 100 rows per page; the loop stops on a short page.
-  // Package-private so the pagination guard can be exercised in tests without 5000 mock rows.
-  static final int CI_PER_PAGE = 100;
-  static final int CI_MAX_PAGES = 50;
 
   private final ThrillhouseConfig config;
   // The login(s) the bot posts under, resolved once from config so that summary dedup, first-review
@@ -112,6 +94,8 @@ public class ReviewOrchestrator {
   private final ReviewDiffFormatter diffFormatter;
 
   private final PrLabeler labeler;
+
+  private final CiStatusEvaluator ciStatusEvaluator;
 
   private final ObjectMapper mapper;
 
@@ -163,6 +147,7 @@ public class ReviewOrchestrator {
       FollowUpAnalyzer followUpAnalyzer,
       ReviewDiffFormatter diffFormatter,
       PrLabeler labeler,
+      CiStatusEvaluator ciStatusEvaluator,
       ObjectMapper mapper) {
     this.config = config;
     this.botIdentity = BotIdentity.from(config.github().botLogins());
@@ -185,6 +170,7 @@ public class ReviewOrchestrator {
     this.followUpAnalyzer = followUpAnalyzer;
     this.diffFormatter = diffFormatter;
     this.labeler = labeler;
+    this.ciStatusEvaluator = ciStatusEvaluator;
     this.mapper = mapper;
   }
 
@@ -329,9 +315,11 @@ public class ReviewOrchestrator {
 
       // Fetch required status checks and evaluate CI checks on the head SHA. An absent result means
       // "could not determine required checks" — evaluateCiChecks then gates on all checks (null).
-      List<String> requiredContexts = resolveRequiredContexts(auth, req).orElse(null);
+      List<String> requiredContexts =
+          ciStatusEvaluator.resolveRequiredContexts(auth, req).orElse(null);
       List<ReviewResult.CiCheck> offendingCiChecks =
-          evaluateCiChecks(auth, req.owner(), req.repo(), req.commitSha(), requiredContexts);
+          ciStatusEvaluator.evaluateCiChecks(
+              auth, req.owner(), req.repo(), req.commitSha(), requiredContexts);
 
       var reviewableFiles = diffFormatter.reviewableFiles(files);
       DiffStats diffStats = DiffStats.fromFiles(reviewableFiles, omittedFiles);
@@ -1365,240 +1353,5 @@ public class ReviewOrchestrator {
               req.commitId(), req.body(), req.event(), List.of());
       reviewClient.createReview(auth, ACCEPT, owner, repo, prNumber, fallback);
     }
-  }
-
-  /**
-   * Resolves the required status-check contexts for the PR's target branch. The contexts come from
-   * two GitHub mechanisms, unioned: repository/organization <em>rulesets</em> (modern; needs only
-   * read access) and <em>classic branch protection</em> (legacy; needs admin). Returns an empty
-   * {@link Optional} — which the caller maps to {@code null}, gating on every check — only when
-   * neither mechanism governs the branch or both lookups fail.
-   */
-  Optional<List<String>> resolveRequiredContexts(String auth, ReviewRequest req) {
-    String branch;
-    try {
-      var prDetails =
-          prClient.getPullRequest(auth, ACCEPT, req.owner(), req.repo(), req.prNumber());
-      if (prDetails == null || prDetails.base() == null || prDetails.base().ref() == null) {
-        return Optional.empty();
-      }
-      branch = prDetails.base().ref();
-    } catch (Exception e) {
-      Log.warnf(e, "Could not resolve target branch; gating CI on all checks");
-      return Optional.empty();
-    }
-
-    var contexts = new LinkedHashSet<String>();
-    boolean resolved = false;
-    var fromRulesets = requiredContextsFromRulesets(auth, req, branch);
-    if (fromRulesets.isPresent()) {
-      resolved = true;
-      contexts.addAll(fromRulesets.get());
-    }
-    var fromClassic = requiredContextsFromClassicProtection(auth, req, branch);
-    if (fromClassic.isPresent()) {
-      resolved = true;
-      contexts.addAll(fromClassic.get());
-    }
-
-    return resolved ? Optional.of(List.copyOf(contexts)) : Optional.empty();
-  }
-
-  /**
-   * Required contexts declared by rulesets governing {@code branch}. An empty {@link Optional}
-   * means no ruleset applies (or the lookup failed); a present-but-empty list means a ruleset
-   * applies but mandates no status checks.
-   */
-  private Optional<List<String>> requiredContextsFromRulesets(
-      String auth, ReviewRequest req, String branch) {
-    try {
-      var rules = checkRunClient.getBranchRules(auth, ACCEPT, req.owner(), req.repo(), branch);
-      if (rules == null || rules.isEmpty()) {
-        return Optional.empty();
-      }
-      var contexts = new ArrayList<String>();
-      for (var rule : rules) {
-        if (rule.isRequiredStatusChecks() && rule.parameters() != null) {
-          for (var check : rule.parameters().requiredStatusChecks()) {
-            if (check.context() != null) {
-              contexts.add(check.context());
-            }
-          }
-        }
-      }
-      return Optional.of(contexts);
-    } catch (Exception e) {
-      Log.warnf(
-          e, "Could not fetch branch rules (rulesets) for %s; trying classic protection", branch);
-      return Optional.empty();
-    }
-  }
-
-  /**
-   * Required contexts declared by classic branch protection. An empty {@link Optional} means the
-   * branch is not protected this way — expected, not an error, for ruleset-only repositories — or
-   * the lookup failed.
-   */
-  private Optional<List<String>> requiredContextsFromClassicProtection(
-      String auth, ReviewRequest req, String branch) {
-    try {
-      var protection =
-          checkRunClient.getRequiredStatusChecks(auth, ACCEPT, req.owner(), req.repo(), branch);
-      return protection == null ? Optional.empty() : Optional.of(protection.contexts());
-    } catch (Exception e) {
-      // A 404 here is normal when the repo uses rulesets instead of classic branch protection.
-      Log.debugf(e, "No classic branch protection for %s", branch);
-      return Optional.empty();
-    }
-  }
-
-  /**
-   * Returns only the <em>offending</em> CI checks on {@code commitSha} — those that are pending,
-   * failing, or (when required) missing entirely. Passing checks are deliberately excluded so the
-   * caller can gate APPROVE on a non-empty result; a successful required check is recorded in
-   * {@code seen} so it is not later mistaken for a missing one.
-   */
-  List<ReviewResult.CiCheck> evaluateCiChecks(
-      String auth, String owner, String repo, String commitSha, List<String> requiredContexts) {
-    var offending = new ArrayList<ReviewResult.CiCheck>();
-    // Required contexts that reported in any state, so the missing-check pass does not re-flag a
-    // green check as pending.
-    var seen = new HashSet<String>();
-    // Contexts already recorded as offending, so a check reported through BOTH the Check Runs API
-    // and the Commit Status API is not listed twice.
-    var offendingNames = new HashSet<String>();
-
-    try {
-      collectPaged(
-          page -> {
-            var resp =
-                checkRunClient.getCheckRuns(
-                    auth, ACCEPT, owner, repo, commitSha, CI_PER_PAGE, page);
-            return resp == null ? null : resp.checkRuns();
-          },
-          run -> addOffendingCheckRun(run, requiredContexts, seen, offendingNames, offending));
-    } catch (Exception e) {
-      Log.warnf(e, "Failed to fetch check runs for commit %s", commitSha);
-    }
-
-    try {
-      collectPaged(
-          page -> {
-            var resp =
-                checkRunClient.getCombinedStatus(
-                    auth, ACCEPT, owner, repo, commitSha, CI_PER_PAGE, page);
-            return resp == null ? null : resp.statuses();
-          },
-          status -> addOffendingStatus(status, requiredContexts, seen, offendingNames, offending));
-    } catch (Exception e) {
-      Log.warnf(e, "Failed to fetch combined status for commit %s", commitSha);
-    }
-
-    addMissingRequiredChecks(requiredContexts, seen, offending);
-    return offending;
-  }
-
-  /**
-   * Pages through a GitHub list endpoint, applying {@code consume} to every row. Stops on an empty
-   * or short page (GitHub's last-page marker) or once {@link #CI_MAX_PAGES} is reached.
-   */
-  private static <T> void collectPaged(IntFunction<List<T>> fetchPage, Consumer<T> consume) {
-    List<T> page = null;
-    for (int p = 1; p <= CI_MAX_PAGES && (page == null || page.size() == CI_PER_PAGE); p++) {
-      page = fetchPage.apply(p);
-      if (page == null) {
-        page = List.of();
-      }
-      page.forEach(consume);
-    }
-  }
-
-  private void addOffendingCheckRun(
-      GitHubCheckRunClient.CheckRunsResponse.CheckRun run,
-      List<String> requiredContexts,
-      Set<String> seen,
-      Set<String> offendingNames,
-      List<ReviewResult.CiCheck> offending) {
-    if (isThrillhouseBotCheck(run.name(), run.app())
-        || isNotRequired(run.name(), requiredContexts)) {
-      return;
-    }
-    seen.add(run.name());
-    String ciStatus = classifyCheckRun(run.status(), run.conclusion());
-    if (!CI_SUCCESS.equals(ciStatus) && offendingNames.add(run.name())) {
-      offending.add(new ReviewResult.CiCheck(run.name(), "check-run", ciStatus, run.conclusion()));
-    }
-  }
-
-  private void addOffendingStatus(
-      GitHubCheckRunClient.CombinedStatus.StatusDetail status,
-      List<String> requiredContexts,
-      Set<String> seen,
-      Set<String> offendingNames,
-      List<ReviewResult.CiCheck> offending) {
-    if (isThrillhouseBotCheck(status.context(), null)
-        || isNotRequired(status.context(), requiredContexts)) {
-      return;
-    }
-    seen.add(status.context());
-    String ciStatus = classifyStatus(status.state());
-    if (!CI_SUCCESS.equals(ciStatus) && offendingNames.add(status.context())) {
-      offending.add(new ReviewResult.CiCheck(status.context(), "status", ciStatus, status.state()));
-    }
-  }
-
-  /** Required contexts that never reported are implicitly pending. */
-  private void addMissingRequiredChecks(
-      List<String> requiredContexts, Set<String> seen, List<ReviewResult.CiCheck> offending) {
-    if (requiredContexts == null) {
-      return;
-    }
-    for (String required : requiredContexts) {
-      // seen.add is false when the context already reported (or is a duplicate entry) — skip those.
-      if (!isThrillhouseBotCheck(required, null) && seen.add(required)) {
-        offending.add(new ReviewResult.CiCheck(required, "missing", CI_PENDING, null));
-      }
-    }
-  }
-
-  private static boolean isNotRequired(String name, List<String> requiredContexts) {
-    return requiredContexts != null && !requiredContexts.contains(name);
-  }
-
-  /** Maps a check-run status/conclusion pair to one of success/pending/failing. */
-  private static String classifyCheckRun(String status, String conclusion) {
-    if (!CHECK_STATUS_COMPLETED.equalsIgnoreCase(status)) {
-      return CI_PENDING;
-    }
-    if (conclusion == null || !PASSING_CONCLUSIONS.contains(conclusion.toLowerCase(Locale.ROOT))) {
-      return CI_FAILING;
-    }
-    return CI_SUCCESS;
-  }
-
-  /** Maps a commit-status state to one of success/pending/failing. */
-  private static String classifyStatus(String state) {
-    if (CI_SUCCESS.equalsIgnoreCase(state)) {
-      return CI_SUCCESS;
-    }
-    if (CONCLUSION_FAILURE.equalsIgnoreCase(state) || "error".equalsIgnoreCase(state)) {
-      return CI_FAILING;
-    }
-    // "pending", null, or any unrecognized state is not a confirmed pass: classify it as pending so
-    // an indeterminate required check holds the approval rather than being mistaken for success
-    // . Only an explicit "success" clears the gate.
-    return CI_PENDING;
-  }
-
-  private boolean isThrillhouseBotCheck(
-      String name, GitHubCheckRunClient.CheckRunsResponse.CheckRun.App app) {
-    if (name != null && (name.equalsIgnoreCase(CHECK_NAME) || containsBotToken(name))) {
-      return true;
-    }
-    return app != null && (containsBotToken(app.slug()) || containsBotToken(app.name()));
-  }
-
-  private static boolean containsBotToken(String value) {
-    return value != null && value.toLowerCase(Locale.ROOT).contains(THRILLHOUSEBOT_TOKEN);
   }
 }
